[feat] 为桌面端应用增加 iframe 加载失败时的提示内容以及重新加载按钮
