+++ apps/desktop/src/index.js
@@ -1,49 +1,8 @@
-// electron/main.js
-
-// 控制应用生命周期和创建原生浏览器窗口的模组
-const {app, BrowserWindow} = require('electron')
-const path = require('path')
-
-const NODE_ENV = process.env.NODE_ENV
-
-function createWindow() {
-    // 创建浏览器窗口
-    const win = new BrowserWindow({
-        width: 1440,
-        height: 900,
-        autoHideMenuBar: true,
-        titleBarStyle: "hidden",
-        titleBarOverlay: {
-            color: 'transparent',
-            symbolColor: '#212121',
-            height: 35
-        },
-        webPreferences: {
-            preload: path.join(__dirname, 'preload.js')
-        }
-    })
-
-    win.loadFile(path.join(__dirname, 'index.html'));
-}
-
-// 这段程序将会在 Electron 结束初始化
-// 和创建浏览器窗口的时候调用
-// 部分 API 在 ready 事件触发后才能使用。
-app.whenReady().then(() => {
-    createWindow()
-
-    app.on('activate', function () {
-        // 通常在 macOS 上，当点击 dock 中的应用程序图标时，如果没有其他
-        // 打开的窗口，那么程序会重新创建一个窗口。
-        if (BrowserWindow.getAllWindows().length === 0) createWindow()
-    })
+const appIframe = document.getElementById('appIframe')
+const appLoadError = document.getElementById('appLoadError')
+const appLoadErrorRefreshButton = document.getElementById('appLoadErrorRefreshButton')
+appIframe.addEventListener('error', () => {
+    appLoadError.dataset.visible = '1'
 })
-
-// 除了 macOS 外，当所有窗口都被关闭的时候退出程序。 因此，通常对程序和它们在
-// 任务栏上的图标来说，应当保持活跃状态，直到用户使用 Cmd + Q 退出。
-app.on('window-all-closed', function () {
-    if (process.platform !== 'darwin') app.quit()
-})
-
-// 在这个文件中，你可以包含应用程序剩余的所有部分的代码，
-// 也可以拆分成几个文件，然后用 require 导入。
+const refreshWindow = () => window.location.reload()
+appLoadErrorRefreshButton.addEventListener('click', refreshWindow)
